+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Added `~/.zshrc.secrets` for work
+- Added `~/.aws/{config,credentials}` for work
+- Use `agenix` for secrets management on macOS
 - Added `bb:`, `bbp:` and `bbro:` Git URL aliases
 - Only use 1 monitor on `phi`
 - Use `wireless` on `phi`